sub-projects/capability-map/scripts/requirements.txt
pyyaml
flask
markdown
jsonschema
